chore(breakpoints): updates documentation template
